test(quant): sort the quant tests by what they cover
Two-level coverage was spread over three files, the same fixture was built four different ways,
and the quantize-then-dequantize launch pair was written out six times, so the file a test lived
in said little about what it checked.

Now: `round_trip.rs` is the sweep over every value type, shape and level that `testgen_quant!`
generates; `two_level.rs` is what only a global scale can get wrong, the binding contract and the
narrow-type cases, with the ue4m3 configuration joining them instead of sitting alone;
`scale_rounding.rs` and `tiled.rs` are unchanged in scope. `harness.rs` holds the two launches, so
a test says what it quantizes rather than how to call the kernel, and `mod.rs` is back to
declarations and the macro.

The block and block-tensor round trips were the same test twice: their scale calibration is the
same formula written two ways, so they differ only in whether the block scales get split by a
global. That is now one function and a flag.

Same 56 tests, same assertions.
