--- crates/cubek-quant/tests/quant/symmetric.rs
@@ -1,324 +1,0 @@
-use cubecl::{
-    features::TypeUsage,
-    ir::ElemType,
-    ir::FloatKind,
-    std::tensor::TensorHandle,
-    {TestRuntime, zspace::shape},
-};
-use cubek_quant::{scheme::QuantMode, scheme::QuantScheme, scheme::QuantStore, scheme::QuantValue};
-
-#[test]
-fn test_quantization_symmetric_tensor() {
-    test_quantization_tensor_symmetric(SHAPE_X, SHAPE_Y, VALUE);
-}
-
-#[test]
-fn test_quantization_symmetric_block() {
-    test_quantization_block_symmetric(
-        SHAPE_X, SHAPE_Y, VALUE, SHAPE_X, // Shape x as block_size
-    );
-}
-
-fn test_quantization_tensor_symmetric(m: usize, n: usize, value: QuantValue) {
-    let mode = QuantMode::Symmetric;
-    let client = TestRuntime::client(&Default::default());
-    let shape = shape![m, n];
-
-    let num_elems: usize = m * n;
-    let half = num_elems as f32 / 2.0;
-    let data: Vec<_> = (0..num_elems).map(|v| v as f32 - half).collect();
-
-    let (q_min, q_max) = value.range();
-    // input data range is not affected by quant range symmetry
-    let scale_f32 = (2.0 * half) / (q_max - q_min);
-
-    let input = f32_tensor(&client, &data, shape.clone());
-    let scale = f32_tensor(&client, &[scale_f32], shape![1]);
-    let output_f = TensorHandle::zeros(&client, shape, f32::as_type_native_unchecked());
-
-    let scheme = QuantScheme::default()
-        .with_level(QuantLevel::Tensor)
-        .with_mode(mode)
-        .with_value(value)
-        .with_store(QuantStore::PackedU32(0))
-        .with_param(QuantParam::F32)
-        .with_mode(QuantMode::Symmetric);
-
-    // The shape is from the POV of packed u32s.
-    let shape_out = shape![m, n / scheme.num_quants()];
-
-    let output = TensorHandle::zeros(&client, shape_out, u32::as_type_native_unchecked());
-    let output_scale = TensorHandle::zeros(&client, shape![1], f32::as_type_native_unchecked());
-
-    cubek_quant::quantize::launch_ref(
-        &client,
-        input.binding(),
-        output.clone().binding(),
-        scale.binding(),
-        None,
-        output_scale.clone().binding(),
-        &scheme,
-        ElemType::Float(FloatKind::F32),
-    )
-    .unwrap();
-
-    cubek_quant::dequantize::launch_ref(
-        &client,
-        // The input of the dequantize kernel is the output of the quantized one.
-        output.binding(),
-        // We use a new buffer to make sure all values are correctly dequantized back.
-        output_f.clone().binding(),
-        output_scale.clone().binding(),
-        None,
-        &scheme,
-        f32::as_type_native_unchecked().storage_type(),
-    )
-    .unwrap();
-
-    let computed = client.read_one_unchecked_tensor(output_f.clone().into_copy_descriptor());
-    let data_restored = f32::from_bytes(&computed);
-
-    // Max quantization error = step size / 2
-    let rel_tol = 1e-4;
-    let max_error = (scale_f32 / 2.0) * (1f32 + rel_tol);
-    assert_eq!(data_restored.len(), data.len());
-    for (actual, expected) in data_restored.iter().zip(data) {
-        let diff = f32::abs(actual - expected);
-        assert!(
-            diff <= max_error,
-            "Expected: {expected} | Actual: {actual} (diff {diff} > {max_error})"
-        );
-    }
-}
-
-fn test_quantization_block_symmetric(m: usize, n: usize, value: QuantValue, block_size: usize) {
-    let mode = QuantMode::Symmetric;
-    let client = TestRuntime::client(&Default::default());
-    let shape = shape![m, n];
-
-    let num_elems: usize = m * n;
-    let half = num_elems as f32 / 2.0;
-    let data: Vec<_> = (0..num_elems)
-        .map(|v| (v as f32 - half) / num_elems as f32)
-        .collect();
-    let (q_min, q_max) = value.range();
-
-    let scale_count = data.len() / block_size;
-    let shape_scale = shape![m, n / block_size];
-
-    let mut scales = Vec::with_capacity(scale_count);
-
-    for block in 0..scale_count {
-        let mut c_max = f32::MIN;
-        let mut c_min = f32::MAX;
-
-        let block_offset = block * block_size;
-
-        for i in 0..block_size {
-            let current = data[block_offset + i];
-            c_max = f32::max(c_max, current);
-            c_min = f32::min(c_min, current);
-        }
-
-        // The bias is assumed to be zero.
-        let range = 2.0 * c_min.abs().max(c_max.abs());
-        let scale = range / (q_max - q_min);
-        scales.push(scale);
-    }
-
-    let input = f32_tensor(&client, &data, shape.clone());
-    let scale = f32_tensor(&client, &scales, shape_scale.clone());
-    let output_f = TensorHandle::zeros(&client, shape, f32::as_type_native_unchecked());
-
-    let scheme = QuantScheme::default()
-        .with_level(QuantLevel::block([block_size as u8]))
-        .with_mode(mode)
-        .with_value(value)
-        .with_store(QuantStore::PackedU32(0))
-        .with_param(QuantParam::F32)
-        .with_mode(QuantMode::Symmetric);
-
-    // The shape is from the POV of packed u32s.
-    let shape_out = shape![m, n / scheme.num_quants()];
-
-    let output = TensorHandle::zeros(&client, shape_out, u32::as_type_native_unchecked());
-    let output_scale = TensorHandle::zeros(&client, shape_scale, f32::as_type_native_unchecked());
-
-    cubek_quant::quantize::launch_ref(
-        &client,
-        input.binding(),
-        output.clone().binding(),
-        scale.binding(),
-        None,
-        output_scale.clone().binding(),
-        &scheme,
-        ElemType::Float(FloatKind::F32),
-    )
-    .unwrap();
-
-    cubek_quant::dequantize::launch_ref(
-        &client,
-        // The input of the dequantize kernel is the output of the quantized one.
-        output.binding(),
-        // We use a new buffer to make sure all values are correctly dequantized back.
-        output_f.clone().binding(),
-        output_scale.binding(),
-        None,
-        &scheme,
-        f32::as_type_native_unchecked().storage_type(),
-    )
-    .unwrap();
-
-    let computed = client.read_one_unchecked_tensor(output_f.clone().into_copy_descriptor());
-    let data_restored = f32::from_bytes(&computed);
-
-    assert_eq!(data_restored.len(), data.len());
-    let rel_tol = 1e-4;
-    for (i, (actual, expected)) in data_restored.iter().zip(data).enumerate() {
-        let block = i / block_size;
-        let scale = scales[block];
-        // Max quantization error = step size / 2
-        let max_error = (scale / 2.0) * (1f32 + rel_tol);
-        let diff = f32::abs(actual - expected);
-        assert!(
-            diff <= max_error,
-            "Mismatch at {i}, Expected: {expected} | Actual: {actual} (diff {diff} > {max_error})"
-        );
-    }
-}
-
-#[test]
-fn test_quantization_symmetric_block_tensor() {
-    // Native storage holds byte-wide values only; the packed sibling covers every value.
-    if !matches!(VALUE, QuantValue::Q8F | QuantValue::Q8S) {
-        return;
-    }
-
-    test_quantization_block_tensor_symmetric(SHAPE_X, SHAPE_Y, VALUE, SHAPE_X, QuantStore::Native);
-}
-
-#[test]
-fn test_quantization_symmetric_block_tensor_packed() {
-    test_quantization_block_tensor_symmetric(
-        SHAPE_X,
-        SHAPE_Y,
-        VALUE,
-        SHAPE_X,
-        QuantStore::PackedU32(0),
-    );
-}
-
-/// The block scales are deliberately split so that neither level alone reconstructs the data, so
-/// dropping the per-tensor scale or applying it twice cannot hide inside the tolerance.
-fn test_quantization_block_tensor_symmetric(
-    m: usize,
-    n: usize,
-    value: QuantValue,
-    block_size: usize,
-    store: QuantStore,
-) {
-    let client = TestRuntime::client(&Default::default());
-    if store == QuantStore::Native && !i8::supported_uses(&client).contains(TypeUsage::Conversion) {
-        return; // backend has no native i8 (e.g. wgpu), which packed storage does not need
-    }
-
-    let shape = shape![m, n];
-
-    let num_elems: usize = m * n;
-    let half = num_elems as f32 / 2.0;
-    let data: Vec<_> = (0..num_elems)
-        .map(|v| (v as f32 - half) / num_elems as f32)
-        .collect();
-    let (q_min, q_max) = value.range();
-    let scale_count = data.len() / block_size;
-    let shape_scale = shape![m, n / block_size];
-
-    let mut raw = Vec::with_capacity(scale_count);
-    for block in 0..scale_count {
-        let mut amax = 0.0f32;
-        for i in 0..block_size {
-            amax = amax.max(data[block * block_size + i].abs());
-        }
-        raw.push(2.0 * amax / (q_max - q_min));
-    }
-
-    let peak = raw.iter().copied().fold(0.0f32, f32::max);
-    let global_f32 = peak / 4.0;
-    let scales: Vec<f32> = raw.iter().map(|s| s / global_f32).collect();
-
-    let input = f32_tensor(&client, &data, shape.clone());
-    let scale = f32_tensor(&client, &scales, shape_scale.clone());
-    let global = f32_tensor(&client, &[global_f32], shape![1]);
-    let output_f = TensorHandle::zeros(&client, shape.clone(), f32::as_type_native_unchecked());
-
-    let scheme = QuantScheme::default()
-        .with_level(QuantLevel::block_tensor(
-            [block_size as u8],
-            QuantParam::F32,
-        ))
-        .with_value(value)
-        .with_store(store)
-        .with_param(QuantParam::F32)
-        .with_mode(QuantMode::Symmetric);
-
-    let output = match store {
-        QuantStore::Native => {
-            TensorHandle::zeros(&client, shape.clone(), i8::as_type_native_unchecked())
-        }
-        // The shape is from the POV of packed u32s.
-        _ => TensorHandle::zeros(
-            &client,
-            shape![m, n / scheme.num_quants()],
-            u32::as_type_native_unchecked(),
-        ),
-    };
-    let output_scale = TensorHandle::zeros(
-        &client,
-        shape_scale.clone(),
-        f32::as_type_native_unchecked(),
-    );
-
-    cubek_quant::quantize::launch_ref(
-        &client,
-        input.binding(),
-        output.clone().binding(),
-        scale.binding(),
-        Some(global.clone().binding()),
-        output_scale.clone().binding(),
-        &scheme,
-        ElemType::Float(FloatKind::F32),
-    )
-    .unwrap();
-
-    cubek_quant::dequantize::launch_ref(
-        &client,
-        output.binding(),
-        output_f.clone().binding(),
-        output_scale.binding(),
-        Some(global.clone().binding()),
-        &scheme,
-        f32::as_type_native_unchecked().storage_type(),
-    )
-    .unwrap();
-
-    let written = client.read_one_unchecked_tensor(global.clone().into_copy_descriptor());
-    assert_eq!(f32::from_bytes(&written)[0], global_f32);
-
-    let computed = client.read_one_unchecked_tensor(output_f.clone().into_copy_descriptor());
-    let data_restored = f32::from_bytes(&computed);
-
-    assert_eq!(data_restored.len(), data.len());
-    for (block, chunk) in data.chunks(block_size).enumerate() {
-        // Error is bounded by half a step of the effective scale, which is the product of the two.
-        let max_error = (global_f32 * scales[block] / 2.0) * (1f32 + 1e-4);
-        for (i, expected) in chunk.iter().enumerate() {
-            let actual = data_restored[block * block_size + i];
-            let diff = f32::abs(actual - expected);
-            assert!(
-                diff <= max_error,
-                "block {block} index {i}: got {actual}, expected {expected}, \
-                 diff {diff} over {max_error}"
-            );
-        }
-    }
-}
